requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
xlrd>=2.0.0
openpyxl>=3.1.0
xlwt>=1.3.0
